Extract relation editor
