+++ src/components/EditableCell.js
@@ -1,5 +1,4 @@
-import { __, sprintf } from '@wordpress/i18n';
-import apiFetch from '@wordpress/api-fetch';
+import { __ } from '@wordpress/i18n';
 import {
 	Button,
 	CheckboxControl,
@@ -7,7 +6,6 @@ import {
 	Dropdown,
 	Notice,
 	Popover,
-	Spinner,
 	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
 	__experimentalNumberControl as NumberControl,
 	TextControl,
@@ -21,16 +19,15 @@ import {
 	useRef,
 	useState,
 } from '@wordpress/element';
-import { Icon, check, closeSmall, plus } from '@wordpress/icons';
+import { Icon, check } from '@wordpress/icons';
 
 import MultiselectEdit from './MultiselectEdit';
 import Chip from './fields/Chip';
 import EditOptionsPopover from './fields/EditOptionsPopover';
 import { resolveFormatColor } from './fields/formatColors';
 import { toRecordId } from '../hooks/fieldIds';
+import RelationEditor from './relations/RelationEditor';
 import RelationReferences from './relations/RelationReferences';
-import { relationIds, relationTitle } from './relations/relationUtils';
-import useCollectionRows from '../hooks/useCollectionRows';
 
 // Resolves the WordPress site locale into a BCP 47 tag for Intl. WP
 // stores locales as `en_US` / `de_DE`; Intl expects `en-US` / `de-DE`.
@@ -98,16 +95,6 @@ function FieldError( { message } ) {
 // http(s) URLs. Anything else (relative paths, mailto:, plain strings) is
 // rendered as text so we never produce a broken link.
 const URL_PATTERN = /^https?:\/\//i;
-const RELATION_PICKER_VIEW = {
-	type: 'table',
-	fields: [],
-	sort: null,
-	filters: [],
-	perPage: 25,
-	page: 1,
-	search: '',
-	layout: {},
-};
 
 // Clamp decimals to a sane range; Intl.NumberFormat throws above 100.
 function clampDecimals( decimals ) {
@@ -629,272 +616,6 @@ function SelectEditor( {
 				</Popover>
 			) : null }
 		</>
-	);
-}
-
-function RelationEditor( { value, relation, onSave, onCancel, label } ) {
-	const [ search, setSearch ] = useState( '' );
-	const [ isCreating, setIsCreating ] = useState( false );
-	const [ createError, setCreateError ] = useState( '' );
-	const searchRef = useRef( null );
-	const selectedIds = useMemo( () => relationIds( value ), [ value ] );
-	const targetCollectionId = Number( relation?.targetCollectionId );
-	const isMultiple = relation?.multiple !== false;
-	const {
-		data,
-		isLoading,
-		refresh: refreshTargetRows,
-	} = useCollectionRows( targetCollectionId || null, RELATION_PICKER_VIEW );
-	const createTitle = search.trim();
-
-	const rows = useMemo( () => {
-		const term = search.trim().toLowerCase();
-		if ( ! term ) {
-			return data;
-		}
-		return data.filter( ( row ) =>
-			( row.title?.raw || row.title?.rendered || '' )
-				.toLowerCase()
-				.includes( term )
-		);
-	}, [ data, search ] );
-	const selectedRefs = useMemo( () => {
-		const currentRefs = Array.isArray( value ) ? value : [ value ];
-		return selectedIds.map(
-			( id ) =>
-				data.find( ( row ) => row.id === id ) ||
-				currentRefs.find( ( ref ) => Number( ref?.id ) === id ) || {
-					id,
-				}
-		);
-	}, [ data, selectedIds, value ] );
-	const unselectedRows = rows.filter(
-		( row ) => ! selectedIds.includes( row.id )
-	);
-	const hasExactMatch =
-		createTitle.length > 0 &&
-		data.some(
-			( row ) =>
-				relationTitle( row ).trim().toLowerCase() ===
-				createTitle.toLowerCase()
-		);
-	const canCreate =
-		targetCollectionId > 0 &&
-		createTitle.length > 0 &&
-		! isLoading &&
-		! hasExactMatch;
-
-	const commit = async ( nextIds ) => {
-		await onSave( nextIds );
-	};
-
-	const toggle = async ( rowId ) => {
-		if ( isMultiple ) {
-			const next = selectedIds.includes( rowId )
-				? selectedIds.filter( ( id ) => id !== rowId )
-				: [ ...selectedIds, rowId ];
-			await commit( next );
-			return false;
-		}
-		await commit( selectedIds.includes( rowId ) ? [] : [ rowId ] );
-		return true;
-	};
-	const remove = async ( rowId ) => {
-		const targetId = Number( rowId );
-		await commit( selectedIds.filter( ( id ) => id !== targetId ) );
-	};
-	const keepSearchFocused = ( event ) => {
-		event.preventDefault();
-	};
-	const createRelatedRow = async ( onClose ) => {
-		if ( ! canCreate || isCreating ) {
-			return;
-		}
-		setIsCreating( true );
-		setCreateError( '' );
-		try {
-			const created = await apiFetch( {
-				path: `/cortext/v1/collections/${ targetCollectionId }/rows`,
-				method: 'POST',
-				data: { title: createTitle },
-			} );
-			const createdId = Number( created?.id );
-			if ( ! createdId ) {
-				throw new Error(
-					__( 'Related row could not be created.', 'cortext' )
-				);
-			}
-			const nextIds = isMultiple
-				? [
-						...selectedIds.filter( ( id ) => id !== createdId ),
-						createdId,
-				  ]
-				: [ createdId ];
-			await commit( nextIds );
-			refreshTargetRows?.();
-			setSearch( '' );
-			if ( ! isMultiple ) {
-				onClose();
-			}
-		} catch ( error ) {
-			setCreateError(
-				error?.message ||
-					__( 'Related row could not be created.', 'cortext' )
-			);
-		} finally {
-			setIsCreating( false );
-		}
-	};
-	useEffect( () => {
-		searchRef.current?.focus();
-	}, [] );
-
-	return (
-		<Dropdown
-			defaultOpen
-			onClose={ onCancel }
-			popoverProps={ {
-				placement: 'bottom-start',
-				className: 'cortext-relation-edit-popover',
-			} }
-			renderToggle={ ( { isOpen, onToggle } ) => (
-				<Button
-					variant="tertiary"
-					className="cortext-relation-edit__toggle"
-					onClick={ onToggle }
-					aria-expanded={ isOpen }
-					aria-label={ label }
-				>
-					{ selectedRefs.length ? (
-						<span className="cortext-relation-edit__toggle-refs">
-							{ selectedRefs.map( ( ref ) => (
-								<span
-									key={ ref.id }
-									className="cortext-relation-edit__toggle-ref"
-								>
-									{ relationTitle( ref ) }
-								</span>
-							) ) }
-						</span>
-					) : (
-						__( 'Select rows…', 'cortext' )
-					) }
-				</Button>
-			) }
-			renderContent={ ( { onClose } ) => (
-				<div className="cortext-relation-edit">
-					<div className="cortext-relation-edit__searchbar">
-						<input
-							ref={ searchRef }
-							className="cortext-relation-edit__search"
-							type="search"
-							value={ search }
-							onChange={ ( event ) =>
-								setSearch( event.target.value )
-							}
-							placeholder={ __(
-								'Search or create a row…',
-								'cortext'
-							) }
-							aria-label={ __( 'Search rows', 'cortext' ) }
-						/>
-					</div>
-					{ selectedIds.length > 0 ? (
-						<div className="cortext-relation-edit__section">
-							<div className="cortext-relation-edit__section-label">
-								{ __( 'Selected', 'cortext' ) }
-							</div>
-							<div className="cortext-relation-edit__selected-list">
-								{ selectedRefs.map( ( ref ) => (
-									<span
-										key={ ref.id }
-										className="cortext-relation-edit__selected-pill"
-									>
-										<span className="cortext-relation-edit__selected-title">
-											{ relationTitle( ref ) }
-										</span>
-										<Button
-											className="cortext-relation-edit__remove"
-											icon={ closeSmall }
-											label={ __(
-												'Remove relation',
-												'cortext'
-											) }
-											onMouseDown={ keepSearchFocused }
-											onClick={ () => remove( ref.id ) }
-										/>
-									</span>
-								) ) }
-							</div>
-						</div>
-					) : null }
-					<div className="cortext-relation-edit__section cortext-relation-edit__section--more">
-						<div className="cortext-relation-edit__section-label">
-							{ __( 'Rows', 'cortext' ) }
-						</div>
-						{ isLoading ? (
-							<div className="cortext-relation-edit__loading">
-								<Spinner />
-							</div>
-						) : null }
-						{ ! isLoading && unselectedRows.length === 0 ? (
-							<div className="cortext-relation-edit__empty">
-								{ createTitle
-									? __( 'No results', 'cortext' )
-									: __( 'No rows', 'cortext' ) }
-							</div>
-						) : null }
-						{ ! isLoading &&
-							unselectedRows.map( ( row ) => (
-								<button
-									key={ row.id }
-									type="button"
-									className="cortext-relation-edit__row"
-									onMouseDown={ keepSearchFocused }
-									onClick={ async () => {
-										const shouldClose = await toggle(
-											row.id
-										);
-										if ( shouldClose ) {
-											onClose();
-										}
-									} }
-								>
-									<span className="cortext-relation-edit__row-title">
-										{ relationTitle( row ) }
-									</span>
-								</button>
-							) ) }
-						{ canCreate ? (
-							<button
-								type="button"
-								className="cortext-relation-edit__row cortext-relation-edit__row--create"
-								onMouseDown={ keepSearchFocused }
-								onClick={ () => createRelatedRow( onClose ) }
-								disabled={ isCreating }
-							>
-								<Icon
-									icon={ plus }
-									className="cortext-relation-edit__row-icon"
-								/>
-								<span className="cortext-relation-edit__row-title">
-									{ sprintf(
-										/* translators: %s: row title */
-										__( 'Create row "%s"', 'cortext' ),
-										createTitle
-									) }
-								</span>
-							</button>
-						) : null }
-						{ createError ? (
-							<div className="cortext-relation-edit__error">
-								{ createError }
-							</div>
-						) : null }
-					</div>
-				</div>
-			) }
-		/>
 	);
 }
